chore: improve to use BigDecimal instead double

+++ DigitalBank/src/main/java/com/joilson/DigitalBank/interfaces/IAccount.java
@@ -3,9 +3,11 @@ package com.joilson.DigitalBank.interfaces;
 
 import com.joilson.DigitalBank.models.Account;
 
+import java.math.BigDecimal;
+
 public interface IAccount {
-    void deposit(double amount);
-    void withdraw(double amount);
-    void transfer(Account toAccount, double amount);
+    void deposit(BigDecimal amount);
+    void withdraw(BigDecimal amount);
+    void transfer(Account toAccount, BigDecimal amount);
     void getStatement();
 }
